[make] Add correct way to create fletch snapshots.
The device needed seems to be called 'the session'. The session is
the baground dart process with a set of settings which point to the
package file.

As of  https://codereview.chromium.org/1408253003/ sod now hosts its
own dart code and packages. Well, one package so far.

BUG=
R=scottmg@chromium.org, travisg@google.com

Review URL: https://codereview.chromium.org/1412803003 .