Add a .clang-format settings file
This is used by "git cl format", upon not finding one it uses LLVM
default format which is not good. This link:

http://clang.llvm.org/docs/ClangFormatStyleOptions.html

Explains the options.

Note that git cl format might not work out of the box. In my case I have
it using the one from the mojo checkout.

BUG=
R=travisg@google.com

Review URL: https://codereview.chromium.org/1412753004 .