[fs][loader] Temporarily remove SPIFS dependencies from SoD
[lk] Note that an LK Roll is also a part of this commit:
    177a43b [dartuino][spifs] Remove target specific fsconfig header.

Since SPIFS does not exist on every platform (QEMU, for example), it
doesn't make sense to have spifs dependencies inside SoD.

Note that this is a stopgap solution as we still have "qspi-flash" and
"spifs" hard coded into the loader. Eventally we'll want the user to
be able to define a full install path (ex. fletch install
/spifs/ex.dart) but for now this should get the build working again
on all platforms.

BUG=
R=cpu@chromium.org

Review URL: https://codereview.chromium.org/1651403002 .